+++ angular-workspace/projects/ni/nimble-angular/src/directives/combobox/tests/nimble-combobox-control-value-accessor.directive.spec.ts
@@ -40,7 +40,7 @@ describe('Nimble combobox control value accessor', () => {
             public selectOptions: (TestModel | null)[] = [
                 { name: 'Option 1', value: 1 },
                 { name: 'Option 2', value: 2 },
-                { name: 'Option 3', value: 3 },
+                { name: ' Option 3 ', value: 3 }, // keep whitespace around name for test
                 null
             ];
 
@@ -94,14 +94,21 @@ describe('Nimble combobox control value accessor', () => {
         });
 
         it('updates selected value when bound property is changed', fakeAsync(() => {
-            testHostComponent.selectedOption = testHostComponent.selectOptions[2];
+            testHostComponent.selectedOption = testHostComponent.selectOptions[0];
             fixture.detectChanges();
             tick();
 
-            expect(combobox.selectedIndex).toBe(2);
+            expect(combobox.selectedIndex).toBe(0);
         }));
 
         it('updates bound property when selected value is changed', () => {
+            clickOnListOption(combobox, 0);
+            fixture.detectChanges();
+
+            expect(testHostComponent.selectedOption).toBe(testHostComponent.selectOptions[0]);
+        });
+
+        it('supports options with whitespace at beginning or end of textContent', () => {
             clickOnListOption(combobox, 2);
             fixture.detectChanges();
 
@@ -237,7 +244,7 @@ describe('Nimble combobox control value accessor', () => {
             public selectOptions: (TestModel | null)[] = [
                 { name: 'Option 1', value: 1 },
                 { name: 'Option 2', value: 2 },
-                { name: 'Option 3', value: 3 },
+                { name: ' Option 3 ', value: 3 }, // keep whitespace around name for test
                 null
             ];
 
@@ -287,14 +294,21 @@ describe('Nimble combobox control value accessor', () => {
         });
 
         it('updates selected value when bound property is changed', fakeAsync(() => {
-            testHostComponent.selectedOption.setValue(testHostComponent.selectOptions[2]);
+            testHostComponent.selectedOption.setValue(testHostComponent.selectOptions[0]);
             fixture.detectChanges();
             tick();
 
-            expect(combobox.selectedIndex).toBe(2);
+            expect(combobox.selectedIndex).toBe(0);
         }));
 
         it('updates bound property when selected value is changed', () => {
+            clickOnListOption(combobox, 0);
+            fixture.detectChanges();
+
+            expect(testHostComponent.selectedOption.value).toBe(testHostComponent.selectOptions[0]);
+        });
+
+        it('supports options with whitespace at beginning or end of textContent', () => {
             clickOnListOption(combobox, 2);
             fixture.detectChanges();