+++ packages/nimble-components/src/anchor-button/tests/anchor-button.spec.ts
@@ -31,22 +31,22 @@ describe('AnchorButton', () => {
 
     it('should set the "control" class on the internal control', async () => {
         await connect();
-        expect(element.control.classList.contains('control')).toBe(true);
+        expect(element.control!.classList.contains('control')).toBe(true);
     });
 
     it('should set the `part` attribute to "control" on the internal control', async () => {
         await connect();
-        expect(element.control.part.contains('control')).toBe(true);
+        expect(element.control!.part.contains('control')).toBe(true);
     });
 
     it('should clear `href` on the internal control when disabled', async () => {
         await connect();
-        element.control.setAttribute('href', 'http://www.ni.com');
+        element.control!.setAttribute('href', 'http://www.ni.com');
 
         element.disabled = true;
         await waitForUpdatesAsync();
 
-        expect(element.control.getAttribute('href')).toBeNull();
+        expect(element.control!.getAttribute('href')).toBeNull();
     });
 
     const attributeNames: { name: string }[] = [
@@ -94,7 +94,7 @@ describe('AnchorButton', () => {
                 element.setAttribute(attribute.name, 'foo');
                 await waitForUpdatesAsync();
 
-                expect(element.control.getAttribute(attribute.name)).toBe(
+                expect(element.control!.getAttribute(attribute.name)).toBe(
                     'foo'
                 );
             });