+++ packages/nimble-components/src/anchor-tree-item/styles.ts
@@ -96,7 +96,7 @@ export const styles = css`
     ${
         /* this rule keeps children without an icon text aligned with parents */ ''
     }
-    span[part="start"] {
+    [part="start"] {
         width: ${iconSize};
     }
 
@@ -113,10 +113,7 @@ export const styles = css`
         height: ${iconSize};
     }
 
-    ${/* the end class is applied when the corresponding slot is filled */ ''}
-    .end {
-        display: flex;
-        fill: currentcolor;
-        margin-inline-start: ${iconSize};
+    [part='end'] {
+        display: none;
     }
 `;