+++ packages/nimble-components/src/label-provider/core/index.ts
@@ -1,0 +1,47 @@
+import { attr } from '@microsoft/fast-element';
+import { DesignSystem } from '@microsoft/fast-foundation';
+import { DesignTokensFor, LabelProviderBase } from '../base';
+import {
+    popupDismissLabel,
+    numericDecrementLabel,
+    numericIncrementLabel
+} from './label-tokens';
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'nimble-label-provider-core': LabelProviderCore;
+    }
+}
+
+const supportedLabels = {
+    popupDismiss: popupDismissLabel,
+    numericDecrement: numericDecrementLabel,
+    numericIncrement: numericIncrementLabel
+} as const;
+
+/**
+ * Core label provider for Nimble
+ */
+export class LabelProviderCore
+    extends LabelProviderBase<typeof supportedLabels>
+    implements DesignTokensFor<typeof supportedLabels> {
+    @attr({ attribute: 'popup-dismiss' })
+    public popupDismiss: string | undefined;
+
+    @attr({ attribute: 'numeric-decrement' })
+    public numericDecrement: string | undefined;
+
+    @attr({ attribute: 'numeric-increment' })
+    public numericIncrement: string | undefined;
+
+    protected override readonly supportedLabels = supportedLabels;
+}
+
+const nimbleLabelProviderCore = LabelProviderCore.compose({
+    baseName: 'label-provider-core'
+});
+
+DesignSystem.getOrCreate()
+    .withPrefix('nimble')
+    .register(nimbleLabelProviderCore());
+export const labelProviderCoreTag = DesignSystem.tagFor(LabelProviderCore);