+++ packages/nimble-components/src/rich-text-editor/index.ts
@@ -1,0 +1,25 @@
+import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
+import { template } from './template';
+import { styles } from './styles';
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'nimble-rich-text-editor': RichTextEditor;
+    }
+}
+
+/**
+ * A nimble styled rich text editor
+ */
+export class RichTextEditor extends FoundationElement {}
+
+const nimbleRichTextEditor = RichTextEditor.compose({
+    baseName: 'rich-text-editor',
+    template,
+    styles
+});
+
+DesignSystem.getOrCreate()
+    .withPrefix('nimble')
+    .register(nimbleRichTextEditor());
+export const richTextEditorTag = DesignSystem.tagFor(RichTextEditor);