+++ packages/nimble-components/src/rich-text-editor/styles.ts
@@ -1,0 +1,12 @@
+import { css } from '@microsoft/fast-element';
+import { display } from '@microsoft/fast-foundation';
+import { bodyFont, bodyFontColor } from '../theme-provider/design-tokens';
+
+export const styles = css`
+    ${display('flex')}
+
+    :host {
+        font: ${bodyFont};
+        color: ${bodyFontColor};
+    }
+`;